requirements.txt
filedate==3.0
mutagen==1.47.0
pillow==12.0.0
requests==2.32.5
